expand sound interfaces with volume getter and setter and apply it to inherited classes. Finilize slider volume settings

+++ src/slidersetting.h
@@ -24,6 +24,6 @@ signals:
 private:
 	QQuickItem* _content;
 	QQmlEngine* _engine;
-	quint64 _value;
+	double _value;
 	int _min, _max;
 };
